fix: repair CI after soroban-sdk 22.0.0 upgrade
Every push/PR has failed all CI checks for the last ~6 days since the
soroban-sdk 22.0.0 dependency bump merged. This fixes the full chain of
fallout:

- Cargo.lock conflict: derive_arbitrary =1.4.2 vs soroban-sdk 22's ~1.3.0
  requirement (drop the unused derive_arbitrary pin)
- compliance-client's compliance dep was dev-only but used unconditionally
  (unresolved import on real builds)
- u8 is no longer TryFromVal-able in soroban-sdk 22; switch compliance
  tier fields to u32
- leftover syntax error in treasury/src/lib.rs from a partial edit
- test suite updated for signature changes (create_invoice, mark_paid,
  treasury.initialize, raise_dispute, block_address) introduced alongside
  the bump
- soroban-sdk 22 testutils no longer accumulates events across separate
  top-level calls, and requires mock_all_auths_allowing_non_root_auth
  for auth checks nested inside cross-contract calls
- fixed a real bug: batch_create_invoice didn't reject duplicate
  merchant nonces within the same batch, only against existing storage
- two stale test assertions/expectations left over from earlier contract
  changes (rotation weight test, cancel-settlement panic message)
- build.yml was missing the required toolchain input for
  dtolnay/rust-toolchain@master (immediate failure before checkout)
- all three contract crates defaulted to the testutils feature, breaking
  release wasm builds; removed the default and added a size-optimized
  release profile so treasury (the largest contract) fits under the
  64KB limit

+++ contracts/compliance/src/lib.rs
@@ -181,7 +181,7 @@ impl ComplianceContract {
         env: Env,
         admin: Address,
         address: Address,
-        tier: u8,
+        tier: u32,
     ) -> Result<(), ContractError> {
         Self::require_admin(&env, &admin)?;
         Self::require_not_paused(&env)?;
@@ -201,11 +201,11 @@ impl ComplianceContract {
     }
 
     /// Returns the stored compliance tier for `address`, or `0` if none has been set.
-    pub fn get_address_tier(env: Env, address: Address) -> u8 {
+    pub fn get_address_tier(env: Env, address: Address) -> u32 {
         env.storage()
             .persistent()
             .get(&DataKey::Tier(address))
-            .unwrap_or(0u8)
+            .unwrap_or(0u32)
     }
 
     pub fn bulk_block_addresses(
@@ -224,7 +224,6 @@ impl ComplianceContract {
         }
         Ok(())
     }
-
 
     // Emergency policy: block_address and clear_address are permitted while paused
     // so the admin can remediate compromised addresses without unpausing first.
@@ -660,9 +659,7 @@ impl ComplianceContract {
             .unwrap_or(Vec::new(env));
         if !index.contains(address) {
             index.push_back(address.clone());
-            env.storage()
-                .instance()
-                .set(&DataKey::AddressIndex, &index);
+            env.storage().instance().set(&DataKey::AddressIndex, &index);
         }
     }
 }
